WIP EXT_mesh_shader mucho problemo

+++ src/main/resources/assets/nvidium/shaders/occlusion/queries/region/fragment.frag
@@ -1,10 +1,11 @@
 #version 460
 #extension GL_ARB_shading_language_include : enable
+#extension GL_ARB_gpu_shader_int64 : require
 #pragma optionNV(unroll all)
 #define UNROLL_LOOP
-#extension GL_NV_gpu_shader5 : require
-#extension GL_NV_bindless_texture : require
-#extension GL_NV_shader_buffer_load : require
+//#extension GL_NV_gpu_shader5 : require
+//#extension GL_NV_bindless_texture : require
+//#extension GL_NV_shader_buffer_load : require
 
 #import <nvidium:occlusion/scene.glsl>
 layout(early_fragment_tests) in;
@@ -14,10 +15,10 @@ layout(location = 0) out vec4 colour;
 void main() {
     uint uid = gl_PrimitiveID*132471+123571;
     colour = vec4(float((uid>>0)&7)/7, float((uid>>3)&7)/7, float((uid>>6)&7)/7, 1.0);
-    regionVisibility[gl_PrimitiveID] = uint8_t(1);
+    regionVisibility[gl_PrimitiveID] = 1;
 }
 #else
 void main() {
-    regionVisibility[gl_PrimitiveID] = uint8_t(1);
+    regionVisibility[gl_PrimitiveID] = 1;
 }
 #endif
